+++ src/chimera_synteny/data/main.js
@@ -142,4 +142,30 @@ function filterEmptyContainersByFamily(inputFamily) {
     }
 }
 
-initDropdowns();
+function mapOrthomyxoSegmentNumbersToNames(){
+    var segmentArray = ["PB2", "PB1", "PA", "HA", "NP", "NA", "M", "NS"]
+    var elms = document.querySelectorAll("div.plotly-graph-div[data-family-taxon=Orthomyxoviridae]")
+
+    ;[...elms].forEach(x => {
+        if (!x.orthomyxoSegmentIsMapped){
+            x.layout.annotations.forEach(y => {
+                var segment_number = parseInt(y.text.split("=")[1])
+                var segment_name = segmentArray[segment_number - 1]
+                y.text = `segment=${segment_name}`
+            })
+
+            x.orthomyxoSegmentIsMapped = true
+
+            Plotly.relayout(x, x.layout)
+        }
+    })
+}
+
+// init
+
+function init(){
+    initDropdowns();
+    mapOrthomyxoSegmentNumbersToNames();
+}
+
+init();