+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [v0.0.2] - 2026-05-08
+
+### Added
+
+None
+
+### Changed
+
+Orthomyxoviridae segment numbers automatically mapped to segment names.
+
+### Fixed
+
+Tolerate empty figlists when grouping data.
+
+
 ## [v0.0.1] - 2026-05-08
 
 ### Added